[PLAT-963] plugin should not send emails if no valid notifs available
